Move HsmManager interface to new security package
